# api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "阅读记忆助手 API (Phase 2 — RAG)",
    "description": "Reading Memory Agent — 基于 pgvector + RAG 的智能阅读记忆恢复。核心功能：上传 TXT → AI 章节总结 + embedding 向量化 → 语义检索 → 无剧透回顾/问答。",
    "version": "2.0.0"
  },
  "servers": [{ "url": "http://localhost:8080", "description": "本地开发" }],
  "tags": [
    { "name": "小说管理" },
    { "name": "阅读进度" },
    { "name": "AI 解析" },
    { "name": "阅读回顾" },
    { "name": "RAG 检索" }
  ],
  "paths": {
    "/api/novels": {
      "post": {
        "tags": ["小说管理"],
        "summary": "上传 TXT 小说",
        "operationId": "uploadNovel",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary", "description": "TXT 格式小说文件" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "上传成功，后台自动触发 AI 解析",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "novel": { "$ref": "#/components/schemas/Novel" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["小说管理"],
        "summary": "获取小说列表",
        "operationId": "listNovels",
        "responses": {
          "200": {
            "description": "小说列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "novels": { "type": "array", "items": { "$ref": "#/components/schemas/Novel" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{id}": {
      "get": {
        "tags": ["小说管理"],
        "summary": "获取小说详情 + 章节 + 进度",
        "operationId": "getNovel",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "小说详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "novel": { "$ref": "#/components/schemas/Novel" },
                    "chapters": { "type": "array", "items": { "$ref": "#/components/schemas/Chapter" } },
                    "progress": { "$ref": "#/components/schemas/ReadingProgress" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{id}/progress": {
      "put": {
        "tags": ["阅读进度"],
        "summary": "更新阅读进度",
        "operationId": "updateProgress",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["current_chapter"],
                "properties": {
                  "current_chapter": { "type": "integer", "minimum": 1, "example": 500 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "进度已更新",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "current_chapter": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{id}/parse": {
      "post": {
        "tags": ["AI 解析"],
        "summary": "触发 AI 解析（总结 + embedding）",
        "description": "异步批量处理：先 AI 总结 → 生成 embedding 向量 → 存入 pgvector。并发 3 章。",
        "operationId": "triggerParse",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "解析已开始",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{id}/recap": {
      "post": {
        "tags": ["阅读回顾"],
        "summary": "生成阅读恢复回顾（RAG 增强版）",
        "description": "使用 pgvector 语义检索最相关章节（而非简单取最近 N 章），生成无剧透回顾。30s 版 + 3min 版。",
        "operationId": "generateRecap",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "回顾内容（Markdown）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "recap": { "type": "string", "description": "Markdown 格式回顾" }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["阅读回顾"],
        "summary": "获取缓存回顾",
        "operationId": "getRecap",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "缓存的回顾",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "recap": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{id}/ask": {
      "post": {
        "tags": ["RAG 检索"],
        "summary": "无剧透智能问答",
        "description": "用户提问 → 问题 Embedding → pgvector 语义检索 → LLM 生成无剧透答案。",
        "operationId": "askQuestion",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["question"],
                "properties": {
                  "question": { "type": "string", "description": "用户问题", "example": "主角现在什么境界？" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI 无剧透回答",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "question": { "type": "string" },
                    "answer": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/novels/{id}/search": {
      "get": {
        "tags": ["RAG 检索"],
        "summary": "语义搜索章节",
        "description": "基于 pgvector 余弦相似度搜索相关章节。自动按阅读进度过滤。",
        "operationId": "searchChapters",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "q", "in": "query", "required": true, "description": "搜索关键词", "schema": { "type": "string", "example": "掌天瓶" } }
        ],
        "responses": {
          "200": {
            "description": "搜索结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": { "type": "string" },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "chapter": { "$ref": "#/components/schemas/Chapter" },
                          "score": { "type": "number", "description": "相似度 (0~1)" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Novel": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string", "example": "凡人修仙传" },
          "total_chapters": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Chapter": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "novel_id": { "type": "integer", "format": "int64" },
          "chapter_number": { "type": "integer" },
          "title": { "type": "string" },
          "summary": { "type": "string", "description": "AI 生成摘要（空=未解析）" },
          "characters": { "type": "array", "items": { "$ref": "#/components/schemas/CharacterInfo" } },
          "events": { "type": "array", "items": { "$ref": "#/components/schemas/EventInfo" } }
        }
      },
      "CharacterInfo": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "aliases": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "string" },
          "first_appearance": { "type": "integer" }
        }
      },
      "EventInfo": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "participants": { "type": "array", "items": { "type": "string" } },
          "summary": { "type": "string" },
          "impact": { "type": "string" },
          "chapter_num": { "type": "integer" }
        }
      },
      "ReadingProgress": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "novel_id": { "type": "integer", "format": "int64" },
          "current_chapter": { "type": "integer" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  }
}